fix leaderboard bug by moving to seperate page from scorepage, tidy up code
