+++ react-client/src/pages/index.js
@@ -2,4 +2,6 @@ export { default as Welcome } from './Welcome';
 export { default as Quiz } from './Quiz';
 export { default as Home } from "./Home";
 export { default as LeaderBoard } from "./LeaderBoard";
+export { default as ScorePage} from "./ScorePage";
+
